Double edges
This commit is part of a larger bug where renames were not handled well.
In this case, renaming a place or transition would not change the name
of the edge. This means that after renaming a place or transition, you
could create a new arc between that place and transition. Even the test
ignored that the name of the edge did not make any logical sense after
renaming one of the connected components.

It now updates the name when one of the connected components is renamed.
It assumes that this rename was valid, thus the new name of the edge
must be unique.