Animation and capacity event
When a Petri net contained more than one type of token, only transitions
where all input places contained all types of tokens were enabled, even
if not all types of tokens were required.

Arcs in PIPE may have an expression as weight instead of a natural
number. It seems an attempt was done to prevent transitions from being
able to fire if this expression evaluated to 0. The code that did this
originally tested if there was a token type where the number of tokens
was -1, likely a signal value of some kind. Later this test was
rewritten to not enable if there was any token type that had a 0 token
count. The code now tests if at least 1 token of any type is removed
from the place, which works better.

Capacity is displayed in PIPE as a tooltip. The capacity that was shown
in the tooltip would not change until something else was changed too.
This was caused by a missing change event when the capacity was changed.
Adding this made the tooltip update automatically.